Test/regexparser

+++ Episode.cs
@@ -2,8 +2,8 @@ using System;
 class Episode
 {
     public string season { get; set; }
-    public int episodeNumberOverall { get; set; } //EpisodeNumber
-    public int episodeNumberInSeason { get; set; } //EpisodeNumber2
+    public float episodeNumberOverall { get; set; } //EpisodeNumber
+    public float episodeNumberInSeason { get; set; } //EpisodeNumber2
     public string title { get; set; }
     public string titleRomaji { get; set; }
     public string titleKanji { get; set; }
